requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.5.0
